flask_service/requirements-flask.txt
flask==3.0.0
flask-cors==5.0.0
psycopg2-binary==2.9.10
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.11.1
pydantic-settings==2.8.1